+++ EmployeeManagementSystem/src/main/java/com/greatlearning/ems/controller/RoleController.java
@@ -13,6 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.greatlearning.ems.dto.RoleDto;
 import com.greatlearning.ems.entity.Role;
+import com.greatlearning.ems.entity.User;
+import com.greatlearning.ems.exception.ResouceNotFoundException;
 import com.greatlearning.ems.spi.RoleService;
 import com.greatlearning.ems.util.RoleHelper;
 
@@ -25,6 +27,13 @@ public class RoleController {
 
 	@PostMapping
 	public String post(@RequestBody() RoleDto role) {
+		
+		var roleByName = roleService.findByName(role.getName());
+
+		if (roleByName.isPresent()) {
+			return String.format("Role %s already Exist", role.getName());
+		}
+		
 		var newRole = new Role(role.getName());
 		
 		if (!RoleHelper.isValid(newRole)) {
@@ -37,13 +46,19 @@ public class RoleController {
 
 	@DeleteMapping("{id}")
 	public String delete(@PathVariable int id) {
+		var resouceById = roleService.findById(id);
+
+		if (resouceById.isEmpty()) {
+			throw new ResouceNotFoundException(Role.class, (long)id);
+		}
+		
 		roleService.deleteById(id);
 		return String.format("Role with id %s Deleted Successfully", id);
 	}
 
 	@GetMapping("{id}")
 	public Role get(@PathVariable int id) {
-		return roleService.findById(id).get();
+		return roleService.findById(id).orElseThrow(() -> new ResouceNotFoundException(Role.class, (long)id));
 	}
 
 	@GetMapping
