+++ EmployeeManagementSystem/src/main/java/com/greatlearning/ems/controller/UserController.java
@@ -12,8 +12,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.greatlearning.ems.dto.UserDto;
+import com.greatlearning.ems.entity.Employee;
 import com.greatlearning.ems.entity.Role;
 import com.greatlearning.ems.entity.User;
+import com.greatlearning.ems.exception.ResouceNotFoundException;
 import com.greatlearning.ems.spi.RoleService;
 import com.greatlearning.ems.spi.UserService;
 import com.greatlearning.ems.util.UserHelper;
@@ -30,6 +32,13 @@ public class UserController {
 
 	@PostMapping()
 	public String post(@RequestBody() UserDto userDto) {
+		
+		var userByUserName = userService.findByUserName(userDto.getUsername());
+
+		if (userByUserName.isPresent()) {
+			return String.format("UserName %s already Exist", userDto.getUsername());
+		}
+		
 		User newUser = new User(userDto.getUsername(), userDto.getPassword(), userDto.getRoles());
 
 		if (!UserHelper.isValid(newUser)) {
@@ -50,13 +59,19 @@ public class UserController {
 
 	@DeleteMapping("{id}")
 	public String delete(@PathVariable int id) {
+		var resouceById = userService.findById(id);
+
+		if (resouceById.isEmpty()) {
+			throw new ResouceNotFoundException(User.class, (long)id);
+		}
+		
 		userService.deleteById(id);
 		return String.format("User with id %s Deleted Successfully", id);
 	}
 
 	@GetMapping("{id}")
 	public User get(@PathVariable int id) {
-		return userService.findById(id).get();
+		return userService.findById(id).orElseThrow(() -> new ResouceNotFoundException(User.class, (long)id));
 	}
 
 	@GetMapping
